Revert "feat: create dev-ci image embedding the frontend (#1726)"
This reverts commit 34ca08c29283832d139567398a63724431cc05ce.

# Conflicts:
#	server/server.go
