+++ Dockerfile
@@ -7,8 +7,6 @@
 
 FROM node:14 as frontend
 
-ARG RELEASE="release"
-
 RUN npm i -g pnpm
 
 WORKDIR /frontend-build
@@ -20,7 +18,7 @@ RUN pnpm install --frozen-lockfile
 COPY ./frontend/ .
 
 # Build frontend
-RUN pnpm "${RELEASE}-docker"
+RUN pnpm release-docker
 
 FROM golang:1.16.5 as backend
 
@@ -30,6 +28,7 @@ ARG GIT_COMMIT="unknown"
 ARG BUILD_TIME="unknown"
 ARG BUILD_USER="unknown"
 
+# Build in release mode so we will embed the frontend
 ARG RELEASE="release"
 
 # Need gcc for CGO_ENABLED=1
